Sigin is a component called in Home, sweetalert don't focus the 'try again' btn, works with normal alert, will try with sweetalert2
